[dynamo] Maintainable code - Move export impl to a different file
[ghstack-poisoned]
